Extract logic out of the views and remove re-base dependency
