+++ package.json
@@ -6,7 +6,6 @@
   "dependencies": {
     "firebase": "^4.6.0",
     "prop-types": "^15.6.0",
-    "re-base": "^3.0.5",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-modal": "^3.0.4",
