+++ src/pages/LandingPage/components/RestaurantManager/index.js
@@ -1,66 +1,27 @@
 import React, { Component } from 'react';
-import base from '../../../../services/rebase';
+import RestaurantRepository from '../../../../services/RestaurantRepository';
 import RestaurantManager from './RestaurantManager';
 
-function generateRestaurantIdFromName(name) {
-  return name
-    .toLowerCase()
-    .replace(/[^a-z0-9-]+/g, '-')
-    .replace(/(^-)?(-$)?/g, '');
-}
-
-function makeNewRestaurantFromName(name) {
-  const id = generateRestaurantIdFromName(name);
-  return {
-    id: id,
-    name: name,
-    createdAt: Date.now(),
-  };
-}
-
 export default class RestaurantManagerContainer extends Component {
-  state = {
-    restaurantList: [],
-  };
+  state = { restaurantList: [] };
+  listener;
 
   componentWillMount() {
-    base.syncState(`restaurants`, {
-      context: this,
-      state: 'restaurantList',
-      asArray: true,
+    this.listener = RestaurantRepository.onChange(list => {
+      this.setState({ restaurantList: list })
     });
   }
 
-  addRestaurant(restaurant) {
-    if (!restaurant) {
-      return false;
-    }
-    const newRestaurant = makeNewRestaurantFromName(restaurant);
-    const idList = this.state.restaurantList.map(restaurant => (restaurant.id));
-
-    if (idList.includes(newRestaurant.id)) {
-      return false;
-    }
-
-    this.setState(({ restaurantList }) => ({
-      restaurantList: [...restaurantList, newRestaurant]
-    }));
-
-    return true;
-  }
-
-  removeRestaurant({ id }) {
-    this.setState(({ restaurantList }) => ({
-      restaurantList: restaurantList.filter(restaurant => restaurant.id !== id)
-    }));
+  componentWillUnmount() {
+    this.listener.off();
   }
 
   render() {
     return (
       <RestaurantManager
         restaurants={this.state.restaurantList}
-        onAdd={restaurant => this.addRestaurant(restaurant)}
-        onRemove={restaurant => this.removeRestaurant(restaurant)}
+        onAdd={restaurant => RestaurantRepository.add({ name: restaurant })}
+        onRemove={({ id }) => RestaurantRepository.remove({ id })}
       />
     )
   }
