+++ src/services/auth/user.js
@@ -1,4 +1,4 @@
-import { app } from '../rebase';
+import { app } from '../firebase';
 import firebase from 'firebase';
 
 export const auth = firebase.auth();
